backend/requirements.txt
azure-servicebus
loguru
pydantic
pydantic-settings
starlette
uvicorn[standard]
